[S12] Rename base package
