--- src/main/java/com/devdemo/app/ws/mobiledemoappws/MobileDemoAppWsApplication.java
@@ -1,27 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@SpringBootApplication
-public class MobileDemoAppWsApplication extends SpringBootServletInitializer {
-
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
-        return applicationBuilder.sources(MobileDemoAppWsApplication.class);
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(MobileDemoAppWsApplication.class, args);
-    }
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
