--- src/main/java/com/devdemo/app/ws/mobiledemoappws/SpringApplicationContext.java
@@ -1,21 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws;
-
-import org.springframework.beans.BeansException;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.stereotype.Component;
-
-@Component
-public class SpringApplicationContext implements ApplicationContextAware {
-	
-	private static ApplicationContext CONTEXT;
-
-	@Override
-	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-		CONTEXT = applicationContext;
-	}
-
-	public static Object getBean(String beanName) {
-		return CONTEXT.getBean(beanName);
-	}
-}
