--- src/main/java/com/devdemo/app/ws/mobiledemoappws/security/AuthenticationFilter.java
@@ -1,69 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws.security;
-
-import com.devdemo.app.ws.mobiledemoappws.SpringApplicationContext;
-import com.devdemo.app.ws.mobiledemoappws.service.UserService;
-import com.devdemo.app.ws.mobiledemoappws.shared.dto.UserDto;
-import com.devdemo.app.ws.mobiledemoappws.ui.model.request.UserLoginRequestModel;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.io.IOException;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Base64;
-import java.util.Date;
-
-public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
-    public AuthenticationFilter(AuthenticationManager authenticationManager) {
-        super(authenticationManager);
-    }
-
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response)
-            throws AuthenticationException {
-        try {
-
-            UserLoginRequestModel creds = new ObjectMapper().readValue(request.getInputStream(), UserLoginRequestModel.class);
-
-            return getAuthenticationManager().authenticate(
-                    new UsernamePasswordAuthenticationToken(creds.getEmail(), creds.getPassword(), new ArrayList<>()));
-
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    protected void successfulAuthentication(HttpServletRequest req, HttpServletResponse res, FilterChain chain,
-                                            Authentication auth) throws IOException, ServletException {
-
-        byte[] secretKeyBytes = Base64.getEncoder().encode(SecurityConstants.getTokenSecret().getBytes());
-        SecretKey secretKey = new SecretKeySpec(secretKeyBytes, SignatureAlgorithm.HS512.getJcaName());
-        Instant now = Instant.now();
-
-        String userName = ((User) auth.getPrincipal()).getUsername();
-        String token = Jwts.builder()
-                .setSubject(userName)
-                .setExpiration(
-                        Date.from(now.plusMillis(SecurityConstants.EXPIRATION_TIME)))
-                .setIssuedAt(Date.from(now)).signWith(secretKey, SignatureAlgorithm.HS512).compact();
-
-        UserService userService = (UserService) SpringApplicationContext.getBean("userServiceImpl");
-        final UserDto userDto = userService.getUser(userName);
-        res.addHeader(SecurityConstants.AUTH_HEADER, SecurityConstants.TOKEN_PREFIX + token);
-        res.addHeader(SecurityConstants.USER_ID_HEADER, userDto.getUserId());
-    }
-}
