--- src/main/java/com/devdemo/app/ws/mobiledemoappws/shared/dto/UserDto.java
@@ -1,19 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws.shared.dto;
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-@Data
-public class UserDto implements Serializable {
-    private static final long serialVersionUID = 1L;
-    private long Id;
-    private String userId;
-    private String firstName;
-    private String lastName;
-    private String password;
-    private String email;
-    private String encryptedPassword;
-    private String emailVerificationToken;
-    private Boolean emailVerified = false;
-}
