--- src/main/java/com/devdemo/app/ws/mobiledemoappws/ui/model/request/UserDetailsRequestModel.java
@@ -1,11 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws.ui.model.request;
-
-import lombok.Data;
-
-@Data
-public class UserDetailsRequestModel {
-        private String firstName;
-        private String lastName;
-        private String email;
-        private String password;
-}
