--- src/main/java/com/devdemo/app/ws/mobiledemoappws/ui/model/request/UserLoginRequestModel.java
@@ -1,9 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws.ui.model.request;
-
-import lombok.Data;
-
-@Data
-public class UserLoginRequestModel {
-    private String email;
-    private String password;
-}
