--- src/main/java/com/devdemo/app/ws/mobiledemoappws/ui/model/response/UserDetailsResponseModel.java
@@ -1,11 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws.ui.model.response;
-
-import lombok.Data;
-
-@Data
-public class UserDetailsResponseModel {
-    private String userId;
-    private String firstName;
-    private String lastName;
-    private String email;
-}
