+++ src/main/java/com/devdemo/app/ws/shared/util/Constant.java
@@ -1,0 +1,7 @@
+package com.devdemo.app.ws.shared.util;
+
+public class Constant {
+    public static final int ENCRYPTED_PASSWORD_LENGTH = 8;
+    public static final int USER_ID_LENGTH = 30;
+
+}
