+++ src/main/java/com/devdemo/app/ws/ui/model/request/UserDetailsRequestModel.java
@@ -1,0 +1,11 @@
+package com.devdemo.app.ws.ui.model.request;
+
+import lombok.Data;
+
+@Data
+public class UserDetailsRequestModel {
+        private String firstName;
+        private String lastName;
+        private String email;
+        private String password;
+}
