+++ src/main/java/com/devdemo/app/ws/ui/model/request/UserLoginRequestModel.java
@@ -1,0 +1,9 @@
+package com.devdemo.app.ws.ui.model.request;
+
+import lombok.Data;
+
+@Data
+public class UserLoginRequestModel {
+    private String email;
+    private String password;
+}
