+++ src/main/java/com/devdemo/app/ws/ui/model/response/UserDetailsResponseModel.java
@@ -1,0 +1,11 @@
+package com.devdemo.app.ws.ui.model.response;
+
+import lombok.Data;
+
+@Data
+public class UserDetailsResponseModel {
+    private String userId;
+    private String firstName;
+    private String lastName;
+    private String email;
+}
