+++ src/test/java/com/devdemo/app/ws/MobileDemoAppWsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.devdemo.app.ws;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MobileDemoAppWsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
