--- src/test/java/com/devdemo/app/ws/mobiledemoappws/MobileDemoAppWsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devdemo.app.ws.mobiledemoappws;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MobileDemoAppWsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
